Remove unused variables and move DomainPresenceInfoUpdateStep into separate class.
