+++ operator/src/main/java/oracle/kubernetes/operator/steps/ServerDownIteratorStep.java
@@ -199,19 +199,4 @@ public class ServerDownIteratorStep extends Step {
     return new DomainPresenceInfoUpdateStep(serverName, next);
   }
 
-  private class DomainPresenceInfoUpdateStep extends Step {
-    public final String serverName;
-
-    public DomainPresenceInfoUpdateStep(String serverName, Step next) {
-      super(next);
-      this.serverName = serverName;
-    }
-
-    @Override
-    public NextAction apply(Packet packet) {
-      DomainPresenceInfo info = packet.getSpi(DomainPresenceInfo.class);
-      info.setServerPod(serverName, null);
-      return doNext(packet);
-    }
-  }
 }
